Collectable class
Simple class for creating and managing collectables. It has an enum to determine what kind of collectable it is and a simple method to determine what to do with it. The money system is not yet finished so I wasn't able to implement the effects but it should be simple to do.